+++ src/test/java/testclasses/exceptions/invoke/invokevirtual/CatchInvokeVirtualObject.java
@@ -4,44 +4,34 @@ import eu.fbk.hardening.annotation.Obfuscation;
 import eu.fbk.hardening.annotation.Protections;
 import testclasses.exceptions.UserDefinedException;
 
-public class CatchInvokeVirtualObject
-{
-    public CatchInvokeVirtualObject()
-    {
+public class CatchInvokeVirtualObject {
+    public CatchInvokeVirtualObject() {
 
     }
 
     @Obfuscation(protections = Protections.TO_NATIVE_CODE)
-    public int div(int a)
-    {
+    public int div(int a) {
         int res = 0;
-        try
-        {
+        try {
             res = ((Integer) normalDiv(a, 0));
-        } catch (ArithmeticException e0)
-        {
-            try
-            {
+        } catch (ArithmeticException e0) {
+            try {
                 other();
                 res += 1000;
-            } catch (UserDefinedException e1)
-            {
+            } catch (UserDefinedException e1) {
                 res++;
-            } catch (Exception e)
-            {
+            } catch (Exception e) {
                 res -= 10000;
             }
         }
         return res;
     }
 
-    public Object normalDiv(int a, int b)
-    {
-        return (Integer) (a / b);
+    public Object normalDiv(int a, int b) {
+        return a / b;
     }
 
-    public void other() throws UserDefinedException
-    {
+    public void other() throws UserDefinedException {
         throw new UserDefinedException();
     }
 }
