+++ src/test/java/testclasses/exceptions/nullpointer/arrayload/CatchNullPointerExceptionXLOADInt.java
@@ -4,51 +4,40 @@ import eu.fbk.hardening.annotation.Obfuscation;
 import eu.fbk.hardening.annotation.Protections;
 import testclasses.exceptions.nullpointer.SupportClassInt;
 
-public class CatchNullPointerExceptionXLOADInt
-{
+public class CatchNullPointerExceptionXLOADInt {
 
-    public CatchNullPointerExceptionXLOADInt()
-    {
+    public CatchNullPointerExceptionXLOADInt() {
 
     }
 
     @Obfuscation(protections = Protections.TO_NATIVE_CODE)
-    public int exec(SupportClassInt obj)
-    {
+    public int exec(SupportClassInt obj) {
         int res = 0;
-        try
-        {
-            res += (int) obj.b[1];
+        try {
+            res += obj.b[1];
             res += 1000;
-        } catch (NullPointerException e)
-        {
+        } catch (NullPointerException e) {
             res++;
         }
 
-        try
-        {
-            res += (int) obj.b[1];
+        try {
+            res += obj.b[1];
             res += 1000;
-        } catch (RuntimeException e)
-        {
+        } catch (RuntimeException e) {
             res++;
         }
 
-        try
-        {
-            res += (int) obj.b[1];
+        try {
+            res += obj.b[1];
             res += 1000;
-        } catch (Exception e)
-        {
+        } catch (Exception e) {
             res++;
         }
 
-        try
-        {
-            res += (int) obj.b[1];
+        try {
+            res += obj.b[1];
             res += 1000;
-        } catch (Throwable e)
-        {
+        } catch (Throwable e) {
             res++;
         }
